cluster/scheduling: refactor leader balancer types to separate file
(cherry picked from commit 0c644d10b256000b78cb73ac0d741759a11ae5d8)
