+++ src/v/cluster/scheduling/leader_balancer_strategy.h
@@ -10,6 +10,7 @@
  */
 #pragma once
 
+#include "cluster/scheduling/leader_balancer_types.h"
 #include "absl/container/btree_map.h"
 #include "absl/container/node_hash_map.h"
 #include "cluster/types.h"
@@ -31,28 +32,19 @@ public:
      * a leader of the group. For convenience, the data structure also contains
      * the replica set of each group.
      */
-    using index_type = absl::node_hash_map<
-      model::broker_shard,
-      absl::btree_map<raft::group_id, std::vector<model::broker_shard>>>;
+    using index_type = leader_balancer_types::index_type;
 
     virtual ~leader_balancer_strategy() = default;
 
     /*
      * Represent leadership transfer for a group.
      */
-    struct reassignment {
-        raft::group_id group;
-        model::broker_shard from;
-        model::broker_shard to;
-    };
+    using reassignment = leader_balancer_types::reassignment;
 
     /*
      * Leaders per shard.
      */
-    struct shard_load {
-        model::broker_shard shard;
-        size_t leaders;
-    };
+    using shard_load = leader_balancer_types::shard_load;
 
     /*
      * Compute error for the current leadership configuration.
